chore: rename cta to slides
